feat: Add parameter to build to let moddir update not delete everything
When it was switched to delete the mods dir, that made it quickly not
useful again for anyone using any additional mods, like the optional
compat mods or clientside mods.

This isn't ideal either, because old versions of updated mods remain in
the mod folder, but sorting mods by last modified in Prism helps
immensely in manual cleanup - every jar intended to be in the base pack
will have their modification time updated to when the script was run, as
they are all overwritten.